Parser: simple strings and simple maps
Maps are very prototypey, cause I'm using autoHashStrat with .deep
setting. So it recursively hashes the field of the struct.

This is not good really because Expressions have position information,
which is now also used in hashing. So if the position changes but the
contents don't it breaks the equality.

Maybe it's ok for my current approach where I run the parsing once and
then show the visualization but if I wanna monitor the changes it
might be useful to know the real difference.