Add a janky label rendering
The UI still misses a layouting system so the positions are specified
from the caller.

The border/backgrounds are a mess. I was changing the screen's stuff
but changed the normal border stuff. This commit changes it back
somewhat.

Need to add padding now that I got the border of a lable bounding
correctly (label should not have borders though, but I'm using them
for debugging).