Add parsing quoted expressions
I'm adding a quoted bool flag to each expression and only set it to
true if there's a quote before it.
I don't really need to compile this code later, so for now this is
more of a meta data that might be useful in the visualization and to
just not crash during parsing of real-life code.